Fix cast problem in migration

+++ projectforge-business/src/main/java/org/projectforge/framework/persistence/database/DatabaseCoreUpdates.java
@@ -226,8 +226,8 @@ public class DatabaseCoreUpdates
       private boolean isImageDataPreviewMissing()
       {
         return databaseUpdateService.doesTableAttributeExist("t_address", "image_data_preview") == false ||
-            ((int) databaseUpdateService.query("select count(*) from t_address where imagedata is not null AND imagedata != ''").get(0).getEntry(0).getValue())
-                > 0
+            ((Long) databaseUpdateService.query("select count(*) from t_address where imagedata is not null AND imagedata != ''").get(0).getEntry(0).getValue())
+                > 0L
                 && databaseUpdateService.query("select pk from t_address where imagedata is not null AND image_data_preview is not null LIMIT 1").size() < 1;
       }
